♻️ (createComponent): Extract `createComponent` into its own file
